Update vscode launch configuration to attach to running server
This chhange makes the nodemon demon startup on debug and closes it after debugging, making the server running on lower resources when running in non-debug mode.

There is currently no way to start a server locally to test without access to my server though, I may look into that later on.

Closes #4