Extract routes to their own file
app.js now has a new ElectionsController class which contains static classes made to execute when a route is hit.

The new router.js file is the one to launch when launching the application. This should be changed to run an index.js instead, buut it was not done for this commit.

Closes #5