Update database logic to call onCreate callback function if empty
This commit makes the get() and execute() method also call the onCreateExecuteCallback function, therefore, if the database is deleted while the app is running, it will run the create script again.

To prevent unneccessary file system check everytime you get the database, if there is already an active sqlite database opened, the file system will not even be queried and the callback will not be done.