Set payload limit to 3MB to accept 2MB photos + extra data
That 3 MB is a little more and may be tweaked to be lowered as the extra data does not take a full MB, but rather maybe a kilobyte. Thing is, the number of candidates is variable, so we don't know how many bytes there will be in a request, whether the picture is in there or not.

Therefore, currently, bumping the limit to 3MB (from a default of 100kb) is good enough.