Implement synchronous operations for most routes
This prevents multiple request being mixed up in the threads, making the election not synched up.

This is usually bad practice, but since this is not huge operations, the performance impact should be big enough.

Also, this sync only applies to paths that requires synchronous operations, the home page for example is async, as well as the whole server is not affected.